Conflicts:
	LivrOnTime/src/main/java/model/GrapheComplet.java
	LivrOnTime/target/classes/model/GrapheComplet.class
	LivrOnTime/target/classes/vue/DessinerPlan.class

+++ LivrOnTime/src/main/java/util/tsp/TemplateTSP.java
@@ -6,14 +6,14 @@ import java.util.Iterator;
 public abstract class TemplateTSP implements TSP {
 	
 	private Integer[] meilleureSolution;
-	private int coutMeilleureSolution = 0;
+	private double coutMeilleureSolution = 0;
 	private Boolean tempsLimiteAtteint;
 	
 	public Boolean getTempsLimiteAtteint(){
 		return tempsLimiteAtteint;
 	}
 	
-	public void chercheSolution(int tpsLimite, int nbSommets, int[][] cout, int[] duree){
+	public void chercheSolution(int tpsLimite, int nbSommets, double[][] cout, int[] duree){
 		tempsLimiteAtteint = false;
 		coutMeilleureSolution = Integer.MAX_VALUE;
 		meilleureSolution = new Integer[nbSommets];
@@ -30,7 +30,7 @@ public abstract class TemplateTSP implements TSP {
 		return meilleureSolution[i];
 	}
 	
-	public int getCoutMeilleureSolution(){
+	public double getCoutMeilleureSolution(){
 		return coutMeilleureSolution;
 	}
 	
@@ -43,7 +43,7 @@ public abstract class TemplateTSP implements TSP {
 	 * @return une borne inferieure du cout des permutations commencant par sommetCourant, 
 	 * contenant chaque sommet de nonVus exactement une fois et terminant par le sommet 0
 	 */
-	protected abstract int bound(Integer sommetCourant, ArrayList<Integer> nonVus, int[][] cout, int[] duree);
+	protected abstract int bound(Integer sommetCourant, ArrayList<Integer> nonVus, double[][] cout, int[] duree);
 	
 	/**
 	 * Methode devant etre redefinie par les sous-classes de TemplateTSP
@@ -53,7 +53,7 @@ public abstract class TemplateTSP implements TSP {
 	 * @param duree : duree[i] = duree pour visiter le sommet i, avec 0 <= i < nbSommets
 	 * @return un iterateur permettant d'iterer sur tous les sommets de nonVus
 	 */
-	protected abstract Iterator<Integer> iterator(Integer sommetCrt, ArrayList<Integer> nonVus, int[][] cout, int[] duree);
+	protected abstract Iterator<Integer> iterator(Integer sommetCrt, ArrayList<Integer> nonVus, double[][] cout, int[] duree);
 	
 	/**
 	 * Methode definissant le patron (template) d'une resolution par separation et evaluation (branch and bound) du TSP
@@ -66,7 +66,7 @@ public abstract class TemplateTSP implements TSP {
 	 * @param tpsDebut : moment ou la resolution a commence
 	 * @param tpsLimite : limite de temps pour la resolution
 	 */	
-	 void branchAndBound(int sommetCrt, ArrayList<Integer> nonVus, ArrayList<Integer> vus, int coutVus, int[][] cout, int[] duree, long tpsDebut, int tpsLimite){
+	 void branchAndBound(int sommetCrt, ArrayList<Integer> nonVus, ArrayList<Integer> vus, double coutVus, double[][] cout, int[] duree, long tpsDebut, int tpsLimite){
 		 if (System.currentTimeMillis() - tpsDebut > tpsLimite){
 			 tempsLimiteAtteint = true;
 			 return;
